+++ R/plotStabilityClass.R
@@ -9,6 +9,10 @@
 #' @param sc The name of the stability class field.
 #' @param type type determines how the data are split and then plotted.
 #' Accepted values are "season" (default) and "hour".
+#' @param locale Locale to use for day and month names. Default is current
+#' locale. Supported locales are listed in stringi::stri_locale_list().
+#' All other labels are in English by default or in Italian if its locale is
+#' specified.
 #'
 #' @return A \code{ggplot2} plot.
 #' 
@@ -18,18 +22,22 @@
 #' 
 #' @export
 #' @examples
-#' data(stMeteo)
+#' data("stMeteo")
 #'
 #' # Season plot of stability class pgt
 #' plotStabilityClass(stMeteo, sc = "pgt", type = "season")
 #' 
 #' # Hourly plot of stability class pgt
 #' plotStabilityClass(stMeteo, sc = "pgt", type = "hour")
+#'
+#' # Override default locale
+#' plotStabilityClass(stMeteo, sc = "pgt", type = "season", locale = "it_IT")
 #' 
-plotStabilityClass <- function(mydata, sc="sc", type="season") {
+plotStabilityClass <- function(mydata, sc="sc", type="season", locale = NULL) {
     
-    if (!requireNamespace("openair", quietly = TRUE)) {
-        stop("Please install openair from CRAN.", call. = FALSE)
+    # Get locale if not explicitely set
+    if (is.null(locale)) {
+        locale <- Sys.getlocale(category = "LC_TIME")
     }
     
     # Fix No visible binding for global variable
@@ -50,16 +58,27 @@ plotStabilityClass <- function(mydata, sc="sc", type="season") {
     mydata$clname <- factor(mydata$clname, levels = sort(unique(mydata$clname),
                                                          decreasing = TRUE))
     
+    if (grepl("it", locale)) {
+        winterLabel <- "Inverno (DGF)"
+        springLabel <- "Primavera (MAM)"
+        summerLabel <- "Estate (GLA)"
+        autumnLabel <- "Autunno (SON)"
+        ylabel <- "Percentuale (%)"
+    } else {
+        winterLabel <- "Winter (DJF)"
+        springLabel <- "Spring (MAM)"
+        summerLabel <- "Summer (JJA)"
+        autumnLabel <- "Autumn (SON)"
+        ylabel <- "Percentage (%)"
+    }
+    
     if (type == "season") {
-        mydata <- openair::cutData(mydata, type = "season")
-        # mydata$quarter <- quarters(mydata$date)
-        # mydata$quarter[mydata$quarter == "Q1"] <- "Inverno"
-        # mydata$quarter[mydata$quarter == "Q2"] <- "Primavera"
-        # mydata$quarter[mydata$quarter == "Q3"] <- "Estate"
-        # mydata$quarter[mydata$quarter == "Q4"] <- "Autunno"
-        # mydata$season <- factor(mydata$season, 
-        # levels = c("Inverno", "Primavera", "Estate", "Autunno"))
-        
+        mydata$month <- as.numeric(format(mydata$date, format = "%m"))
+        mydata$season <- season(mydata$month)
+        mydata$season[mydata$season == 1] <- winterLabel
+        mydata$season[mydata$season == 2] <- springLabel
+        mydata$season[mydata$season == 3] <- summerLabel
+        mydata$season[mydata$season == 4] <- autumnLabel
         mydata$season <- factor(mydata$season, levels = unique(mydata$season))
 
         v <- ggplot(mydata, aes(x = season, fill = clname)) +
@@ -73,7 +92,7 @@ plotStabilityClass <- function(mydata, sc="sc", type="season") {
     v <- v + 
         scale_y_continuous(labels = scales::label_percent(), breaks = seq(0,1,0.1)) +
         scale_fill_brewer(palette = "Spectral", direction = -1) + 
-        labs(x = "", y = "Percentuale (%)") +
+        labs(x = "", y = ylabel) +
         theme_bw(base_family = "sans") +
         theme(legend.position = "bottom") +
         guides(fill = guide_legend(
@@ -85,4 +104,23 @@ plotStabilityClass <- function(mydata, sc="sc", type="season") {
             reverse = TRUE)) 
     
     return(v)
+}
+
+
+season <- function(x) {
+    res <- lapply(lubridate::month(x), function(x) {
+        # Winter
+        if (x %in% c(1, 2, 12)) s = 1
+        # Spring
+        else if (x %in% c(3, 4, 5)) s = 2
+        # Summer
+        else if (x %in% c(6, 7, 8)) s = 3
+        # Autumn
+        else if (x %in% c(9, 10, 11)) s = 4
+        else s = NULL
+        s
+    })
+    res <- unlist(res)
+    
+    return(res)
 }